requirements.txt
telethon==1.30
python-dotenv==1.0.1